Revert "refactor Aggregate root (Project)로부터 Initiative 분리"
This reverts commit 051b68acd54083c770ba7d0655a207837b54ef33.
