--- src/main/java/kr/jay/okrver3/interfaces/initiative/InitiativeApiController.java
@@ -1,14 +1,0 @@
-package kr.jay.okrver3.interfaces.initiative;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-@RestController
-@RequestMapping("/api/v1/initiative")
-@RequiredArgsConstructor
-public class InitiativeApiController {
-}
