--- src/test/java/kr/jay/okrver3/interfaces/initiative/InitiativeApiControllerAcceptanceTest.java
@@ -1,88 +1,0 @@
-package kr.jay.okrver3.interfaces.initiative;
-
-import static org.assertj.core.api.AssertionsForClassTypes.*;
-
-import java.sql.Connection;
-import java.util.regex.Pattern;
-
-import javax.sql.DataSource;
-
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.http.HttpStatus;
-import org.springframework.jdbc.datasource.init.ScriptUtils;
-import org.springframework.transaction.annotation.Transactional;
-
-import io.restassured.RestAssured;
-import io.restassured.http.ContentType;
-import kr.jay.okrver3.TestHelpUtils;
-import kr.jay.okrver3.common.utils.JwtTokenUtils;
-import kr.jay.okrver3.interfaces.project.ProjectInitiativeSaveDto;
-
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-@Transactional
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class InitiativeApiControllerAcceptanceTest {
-
-	@Value("${app.auth.tokenSecret}")
-	private String key;
-
-	@Value("${app.auth.tokenExpiry}")
-	private Long accessExpiredTimeMs;
-
-	@Autowired
-	DataSource dataSource;
-	private static final String baseUrl = "/api/v1/initiative";
-	@LocalServerPort
-	private int port;
-
-	private String authToken;
-
-	@BeforeAll
-	void setUpAll() {
-		try (Connection conn = dataSource.getConnection()) {
-			authToken = JwtTokenUtils.generateToken("initiativeRetrieveTest@naver.com", key, accessExpiredTimeMs);
-			ScriptUtils.executeSqlScript(conn, new ClassPathResource("/insert-project-data.sql"));
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-
-	@Test
-	void 행동전략_추가시_기대하는_응답을_리턴한다_initiativeToken() throws Exception {
-
-		ProjectInitiativeSaveDto requestDto = new ProjectInitiativeSaveDto(
-			"key_wV6MX15WQ3DTzQMs",
-			"행동전략",
-			TestHelpUtils.getDateString(10, "yyyy-MM-dd"),
-			TestHelpUtils.getDateString(-100, "yyyy-MM-dd"),
-			"행동전략 상세내용"
-		);
-
-		final String response = RestAssured.
-
-			given()
-			.contentType(ContentType.JSON)
-			.header("Authorization", "Bearer " + authToken)
-			.body(requestDto).
-
-			when()
-			.post(baseUrl).
-
-			then()
-			.statusCode(HttpStatus.CREATED.value())
-			.extract().body().asString();
-
-		assertThat(response).containsPattern(
-			Pattern.compile("initiative-[a-zA-Z0-9]{10}"));
-	}
-}
