+++ src/test/java/kr/jay/okrver3/interfaces/project/ProjectApiControllerTest.java
@@ -21,6 +21,7 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
 
+import kr.jay.okrver3.TestHelpUtils;
 import kr.jay.okrver3.common.exception.ErrorCode;
 import kr.jay.okrver3.common.exception.OkrApplicationException;
 import kr.jay.okrver3.domain.project.ProjectType;
@@ -274,6 +275,7 @@ class ProjectApiControllerTest {
 
 	}
 
+
 	@Test
 	@Sql("classpath:insert-project-date.sql")
 	void 프로젝트_핵심결과_추가시_기대하는_응답을_리턴한다_keyResultToken() throws Exception {
@@ -287,11 +289,40 @@ class ProjectApiControllerTest {
 		UsernamePasswordAuthenticationToken auth = new UsernamePasswordAuthenticationToken(
 			user, null, user.getAuthorities());
 
-		ResponseEntity<String> response = sut.registerKeyResult(
-			new ProjectKeyResultSaveDto(projectToken, keyResultName), auth);
+		ResponseEntity<String> response = sut.registerKeyResult(new ProjectKeyResultSaveDto(projectToken, keyResultName), auth);
+
 
 		assertThat(response.getBody()).containsPattern(
 			Pattern.compile("keyResult-[a-zA-Z0-9]{10}"));
 
 	}
+
+
+	@Test
+	@Sql("classpath:insert-project-date.sql")
+	void 행동전략_추가시_기대하는_응답을_리턴한다_initiativeToken() throws Exception {
+
+		ProjectInitiativeSaveDto requestDto = new ProjectInitiativeSaveDto(
+			"key_wV6MX15WQ3DTzQMs",
+			"행동전략",
+			TestHelpUtils.getDateString(10, "yyyy-MM-dd"),
+			TestHelpUtils.getDateString(-100, "yyyy-MM-dd"),
+			"행동전략 상세내용"
+		);
+
+		User user = em.createQuery("select u from User u where u.id = :userSeq", User.class)
+			.setParameter("userSeq", 3L)
+			.getSingleResult();
+
+		UsernamePasswordAuthenticationToken auth = new UsernamePasswordAuthenticationToken(
+			user, null, user.getAuthorities());
+
+		ResponseEntity<String> response = sut.registerInitiative(requestDto, auth);
+
+
+		assertThat(response.getBody()).containsPattern(
+			Pattern.compile("initiative-[a-zA-Z0-9]{10}"));
+	}
+
+
 }
